# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Simulação",
        "description": "Documentação da API de Simulação. Esta API fornece funcionalidades para gerenciar clientes, instituições, modalidades de crédito e ofertas.",
        "contact": {
            "name": "Luiz Santos Full Stack Developer ",
            "email": "luizcsdev@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar todos os clientes",
                "description": "Retorna uma lista de clientes cadastrados.",
                "operationId": "ce1b06de9bbd15a5cc0ed7ab28ee5713",
                "responses": {
                    "200": {
                        "description": "Lista de clientes.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Cliente"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhum cliente encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Não há clientes cadastrados."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Cria um novo cliente",
                "description": "Cadastra um novo cliente no sistema.",
                "operationId": "703b9106594ca621e0c5bb1b8328bb70",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "cpf"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "João da Silva"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "123.456.789-00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente criado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente criado com sucesso."
                                        },
                                        "cliente": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nome": {
                                                    "type": "string",
                                                    "example": "João da Silva"
                                                },
                                                "cpf": {
                                                    "type": "string",
                                                    "example": "123.456.789-00"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-12-21T10:00:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-12-21T10:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Os dados fornecidos são inválidos."
                                        },
                                        "errors": {
                                            "properties": {
                                                "nome": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "O campo nome é obrigatório."
                                                    }
                                                },
                                                "cpf": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "O campo cpf é obrigatório."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/clientes/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Exibe os detalhes de um cliente",
                "description": "Retorna os detalhes de um cliente específico, incluindo suas associações com instituições e modalidades.",
                "operationId": "9414678b6b3a1f03b3d296d4dcf8d844",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do cliente retornados com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "João da Silva"
                                        },
                                        "cpf": {
                                            "type": "string",
                                            "example": "123.456.789-00"
                                        },
                                        "instituicoes": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "modalidades": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente não encontrado."
                    }
                }
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Atualiza as informações de um cliente",
                "description": "Atualiza os dados de um cliente existente.",
                "operationId": "27af8d44cc1dbf176f91e6118f0ce961",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "João da Silva"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "123.456.789-00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente atualizado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente atualizado com sucesso."
                                        },
                                        "cliente": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nome": {
                                                    "type": "string",
                                                    "example": "João da Silva"
                                                },
                                                "cpf": {
                                                    "type": "string",
                                                    "example": "123.456.789-00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente não encontrado."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Remove um cliente",
                "description": "Remove um cliente e suas associações com instituições e modalidades, se existirem.",
                "operationId": "9ad82979f62d7065a69ed5067066d569",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente removido com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente removido com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao remover o cliente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao remover o cliente."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Mensagem de erro detalhada."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Cliente": {
                "title": "Cliente",
                "required": [
                    "nome",
                    "cpf"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nome": {
                        "type": "string",
                        "example": "João Silva"
                    },
                    "cpf": {
                        "type": "string",
                        "example": "123.456.789-00"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}